Separate linux specific Refresh method
The UnixProcess seems to be generic enough, except for the Refresh
method, that contains linux specific stuff. Now it should be able to
support everything proc based.
